Move classe ColaBoraProject para serviço de Projetos
